add webpack config
add src files to webpack folder
