+++ webpack/src/scss/libs/_vars.scss
@@ -51,5 +51,5 @@ $palette: (
 
 //pictures startpage
 $picture-startpage: (
-  scroll-1: url(../images/home/abschnitt1.jpeg),
+  scroll-1: url(/../images/home/abschnitt1.jpeg),
 );
