+++ webpack/src/scss/main.scss
@@ -1,8 +1,8 @@
 @import 'libs/vars';
 @import 'libs/functions';
 @import 'libs/mixins';
-@import url("../css/font-awesome.min.css");
-@import url("../fonts/roboto.css");
+@import url("/../css/font-awesome.min.css");
+@import url("/../fonts/roboto.css");
 @import url("https://toolbox-bodensee.de/css/spacestatus.css");
 @import "libs/skel";
 @import 'parts/startpage';
