+++ webpack/src/scss/parts/_banner.scss
@@ -2,7 +2,7 @@
 #banner {
   background-attachment: fixed;
   background-color: _palette(accent2);
-  background-image: url('../../images/banner.jpg');
+  background-image: url('/../../images/banner.jpg');
   background-position: center center;
   background-size: cover;
   box-shadow: 0 0.25em 0.5em 0 rgba(0, 0, 0, 0.25);
@@ -129,4 +129,3 @@ body.is-touch {
     background-attachment: scroll;
   }
 }
-
