+++ webpack/src/scss/parts/_footer.scss
