+++ webpack/src/scss/parts/_icon.scss
