+++ webpack/src/scss/parts/_image.scss
