+++ webpack/src/scss/parts/_section-article.scss
