+++ webpack/src/scss/parts/_spotlight.scss
