+++ webpack/src/scss/parts/_table.scss
